requirements.txt
streamlit>=1.46.0
openai>=1.0.0
python-dotenv>=1.0.0
google-generativeai>=0.8.0
